made cleaning the build folder an explicit choice (flag) instead of hidden feature

+++ scripts/templates.py
@@ -86,12 +86,18 @@ def read_yaml_files(yaml_dirs, name: str = "config.yaml"):
     return data
 
 
-def process_config_files(dirs, yaml_data, build_dir="./build"):
+def process_config_files(dirs : [str], yaml_data: str, build_dir : str ="./build", clean : bool = False):
     
     path = Path(build_dir)
     if path.exists():
-        shutil.rmtree(path)
-    path.mkdir()
+        if any(path.iterdir()):
+            if clean:
+                shutil.rmtree(path)
+                path.mkdir()
+            else:
+                print("Warning: Build directory is not empty. Use --clean to remove existing files.")
+    else:
+        path.mkdir()
 
     local_yaml_path = ""
     yaml_local = {}
@@ -148,10 +154,11 @@ def process_config_files(dirs, yaml_data, build_dir="./build"):
                     f.write(content)
 
 
-def build_config(dir: str = "./build"):
+def build_config(dir: str = "./build", clean: bool = False):
     modules = read_module_config(root_dir="./", tmpl_dirs=TMPL_DIRS)
     process_config_files(
         dirs=modules,
         yaml_data=read_yaml_files(yaml_dirs=YAML_DIRS),
         build_dir=dir,
+        clean = clean,
     )
